requirements.txt
cartoframes
catboost
dash-bootstrap-components
dash-extensions
dash-leaflet
flask-jwt
fuzzywuzzy
geocoder
geopy
googlemaps
imbalanced-learn
jsbeautifier
leafmap
loguru
mpu
openpyxl
overpy
pgeocode
pip-chill
pipwin
plotly-express
pony
psycopg2
rtree
scikit-posthocs
sklearn
streamlit-folium
turfpy
uszipcode
xgboost
yellowbrick
smart_open
pandas
